ui: Move reply test into a separate module
